requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
aophelper==0.3.0
asgiref==3.8.1
astunparse==1.6.3
async-timeout==5.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
elastic-transport==8.17.0
elasticsearch==8.17.1
elasticsearch-dsl==8.17.1
exceptiongroup==1.2.2
fastapi==0.115.8
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
httptools==0.6.4
idna==3.10
imageio==2.37.0
joblib==1.4.2
keras==3.8.0
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mysql-connector-python==9.2.0
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydicom==3.0.1
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gdcm==3.0.24.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
starlette==0.45.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
tifffile==2025.2.18
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
